Delete CommitteUtil, move remaining methods to BeaconStateUtil
